JDBC练习

+++ spirng1javaConfig/src/main/java/com/wil/service/UserService.java
@@ -5,5 +5,5 @@ package com.wil.service;
  */
 public interface UserService {
     void save();
-    int count();
+    Long count();
 }
